backend/requirements.txt
fastapi
uvicorn
python-multipart
pdf2image
pytesseract
pillow
google-genai
python-dotenv
sqlalchemy
sentence-transformers
torch
numpy
requests
opencv-python
easyocr
reportlab
openai
